# Provide a few optimizations to the `matchindex` C++ function: