--- a/man/matchindex.Rd
+++ b/man/matchindex.Rd
@@ -89,6 +89,7 @@ train$mpg[idx]
 # twelve mpg entries in test, and count how many times the true value
 # (which we know here) is located within the inter-quartile range of each
 # distribution. Is your count anywhere close to 500? Why? Why not?
+
 }
 \author{
 Stef van Buuren, Nasinski Maciej, Alexander Robitzsch